+++ Tests/Builds.tests.ps1
@@ -11,21 +11,15 @@ Import-Module -Name (Join-Path $PSScriptRoot '..\VSTS.psm1') -Force
 
 Describe 'Builds' -Tags 'Integration' {
 	BeforeAll {
-		$Script:session = New-VSTSSession -AccountName $account -User $userName -Token $token
-		$Script:projectName = New-ProjectName
-
-		$parameterDetails = @{
-			Session = $Script:session
-			Name    = $Script:projectName
-			Verbose = $True
-		}
-
-		New-VSTSProject @parameterDetails -Wait
+		$projectName = New-ProjectName
+		$session = New-VSTSSession -AccountName $account -User $userName -Token $token
+		Write-Verbose -Verbose -Message ('Creating VSTS test project {0}' -f $projectName)
+		New-VSTSProject -Session $session -Name $projectName -Wait
 	}
 
 	Context 'Has default build queues' {
 		It 'Should return default build queues' {
-			{ $script:queue = Get-VstsBuildQueue -Session $Script:session } | Should Not Throw
+			{ $script:queue = Get-VstsBuildQueue -Session $session } | Should Not Throw
 			$script:queue | Where-Object -Property Name -EQ 'Default' | Should Not BeNullOrEmpty
 			$script:queue | Where-Object -Property Name -EQ 'Hosted' | Should Not BeNullOrEmpty
 			$script:queue | Where-Object -Property Name -EQ 'Hosted Linux Preview' | Should Not BeNullOrEmpty
@@ -34,6 +28,7 @@ Describe 'Builds' -Tags 'Integration' {
 	}
 
 	AfterAll {
-		Remove-VSTSProject @parameterDetails
+		Write-Verbose -Verbose -Message ('Deleting VSTS test project {0}' -f $projectName)
+		Remove-VSTSProject -Session $session -Name $projectName
 	}
 }